1.2.2 patch
- Fixed: A crash with color element due to Integer.decode not supporting
the full Integer range... WHY JAVA

+++ src/main/java/carbonconfiglib/gui/config/ColorElement.java
@@ -66,12 +66,12 @@ public class ColorElement extends ConfigElement
 		super.render(graphics, x, top, left, width, height, mouseX, mouseY, selected, partialTicks);
 		if(isArray()) {
 			graphics.fill(left+186, top-1, left+203, top+19, 0xFFA0A0A0);
-			graphics.fill(left+187, top, left+202, top+18, Integer.decode(value.get()) | 0xFF000000);
+			graphics.fill(left+187, top, left+202, top+18, Long.decode(value.get()).intValue() | 0xFF000000);
 		}
 		else {
 			int xOff = isCompound() ? 194 : 207;
 			graphics.fill(left+xOff, top-1, left+xOff+17, top+19, 0xFFA0A0A0);
-			graphics.fill(left+xOff+1, top, left+xOff+16, top+18, Integer.decode(value.get()) | 0xFF000000);
+			graphics.fill(left+xOff+1, top, left+xOff+16, top+18, Long.decode(value.get()).intValue() | 0xFF000000);
 		}
 		if(textBox.isMouseOver(mouseX, mouseY) && result != null && !result.getValue()) {
 			owner.addTooltips(Component.literal(result.getError().getMessage()).withStyle(ChatFormatting.RED));
